Create D. Segment Intersections
https://codeforces.com/contest/1389/problem/D

Let's first consider n = 1. A solution for this case will inform a case for more general n.

Observation 1: If (l1, r1) and (l2, r2) are disjoint and don't share any length, we'll have to spend a few initial operations to connect them. This is stage 1. It doesn’t matter how we connect them as long as they are connected. Note that this won't contribute to the intersection. Once they are connected, we can increase the total intersection by 1 per move by extending one of the intervals at a time such that they reach the initial range of the two intervals. For example: If the 2 intervals were initially (2, 5) and (7, 9) and we spent 2 moves to connect them (EX: (2, 6), (7, 9)) then we can extend either the first or second interval until both become (2, 9). This stage gives us the best increase in intersection sum per move. In the third stage, it will take 2 moves to increase the intersection sum by 1. Just extend the first interval to the right by 1 and the second interval to the right by 1. Note that there is no limit on this operation, so we are always guaranteed to eventually reach a total intersection of k. 

Observation 2: If (l1, r1) and (l2, r2) aren’t disjoint and share some length, than we can skip stage 1 since they are already connected. Then for stage 2, we will still extend both intervals to the original range of their union. However we will have to account for the fact that there was already an intersection, thus we may be overcounting by just adding the range of their union to the answer. Stage 3 is the same as in the disjoint case.

Now here’s the main solution where we have n copies of the pair of the intervals.

Observation 3: Let’s say that at the end of our process, only k copies of the pair are used and have some moves applied to them. Then they must have all been connected. Proof by contradiction: If a copy hadn’t been connected, then it wouldn’t have contributed to the total intersection sum and shouldn’t have been used in the first place to minimize the total #of moves. However this contradicts the fact that it was used and had moves applied to it. Thus a contradiction exists and it should never have been part of our set of used copies in the first place.

SOLUTION: 
	If the total intersection of all n copies of the pair of segments is >= k, then we don’t need to expend any moves. Otherwise iterate over the #of copies of the pair used. This # has to be less than or equal to n. We will expend a few moves to connect each pair if it wasn’t already connected. Then conduct stage 2 for the n pairs. If the total intersection sum reaches k in this stage, then we have a solution. Otherwise, we will move on to stage 3 and extend one pair of segments as much as possible until our sum reaches k. Pick the smallest total #of moves over all #copies of pairs used as our answer.