Create D. Rarity and New Dress
https://codeforces.com/contest/1393/problem/D

Observation 1: Note that we’re being asked to count the #of “diamonds” such that every element in a valid diamond shares the same color and isn’t out of bounds. 

Algorithm: Let’s iterate over the center point of a diamond. Note that 4 “triangles” are formed, the top-right triangle, the top-left triangle, the bottom-right triangle, the bottom-left triangle. Note that these triangles are overlapping. If for each triangle, we can find the largest length of it, then the #of valid diamonds for a center point is just the minimum of the values of all 4 types of triangles centered at it. Let’s focus on the top-left triangle; the other triangles can be solved with a similar idea – just with different loop orders. Iterate from the top to the bottom and from the left to the right. Let dp[i][j] be the largest length of a top-left triangle centered at (i, j). If i == 0 or j == 0, then we can’t extend the diamond and dp[i][j] = 1. Otherwise if grid[i][j] isn’t equal to grid[i-1][j] or grid[i][j-1], then we also can’t extend the diamond and dp[i][j] = 1. Otherwise, we can set dp[i][j] = min(dp[i-1][j], dp[i][j-1]) + 1. This is easy to see by drawing out some examples. Once we’ve computed our dp for all the top-left triangles, run a similar idea for the other 3 types of triangles. You will have to change the loop order and some of the conditions involving hitting the boundary and the transitions to reflect the nature of the order in which cells are being computed. For example with a bottom-right triangle, dp[i][j] = 1 if i == n-1 or j == m-1.

Finally once all 4 triangles have been computed for every cell, we calculate the #of diamonds centered at each cell by taking the min of all the values of the 4 triangles at that cell and sum this count for all the cells.