Create E. Solo mid Oracle
https://codeforces.com/contest/1413/problem/E

Solution: 

Let’s first consider the case where the operations don’t overlap with each other. Then if the total sum of the operation is negative, then Naruto can keep applying this operation forever and can keep decreasing the health of a monster. Thus he can kill any monster in this case so print -1. Otherwise the highest health monster that he can kill is just the amount of damage that he deals at the beginning of the first operation, since the monster will gain health after that.

Now consider the case where the operations overlap with each other. Note that if we want to maximize the health of a potential enemy while still being able to kill it, we will only consider all units of time from time 0 (where the first move starts) to some time kd (where the k+1 th move starts). This is b/c after that the monster will eventually be able to heal until it reaches the k+2th move, but that’s a contradiction since we already claimed that we will end once some arbitrary move starts). So let’s define prefix i to be from time 0 to time ki inclusive. As an example, prefix 5 would be [0, 5d].

Now notice that the first few prefixes will all be decreasing, while the last few prefixes will all be increasing. This is b/c at the beginning, there aren’t as many operations so there will be less overlap and fewer chances for the monster to heal; therefore its health will decrease. However as more operations are used, the monster will have more chances to heal and its health will start to increase. So we’ll use ternary search on the optimal prefix to use such that we can reduce the monster’s health as much as possible and take that as our answer. We can implement ternary search with binary search by finding the first prefix i such that the largest possible absolute damage you can do to a monster at the end of prefix i is >= the largest possible absolute damage you can do to a monster at the end of prefix i.

To compute the amount of damage we can do to a monster at the end of prefix i, do the following steps. Firstly, the number of operations is just i+1. This is by definition of prefix i. Thus the total #of health we deal the monster is just -1 * (i+1) * a. We also have c * (i+1) healing units of time. The total health that these units of time will save is c * b * (i+1). However we’ve overcounted b/c some of these healing units of time occur after time d * i. So we’ll subtract the effect of those healing units of time. Since the end of prefix i is defined to be time d * i, we know that none of the c healing units of time for operation i+1 should be counted in our final sum. Since the operations are repeated every d units of time, we know that for operation i, c - d healing units of time won’t be counted. For operation i-1, c - 2d healing units of time won’t be counted, and so on. 

Note that the #of extra units of time will look something c + (c-d) + (c-2d) + . . . We can see that we have ceil(c/d) terms. Let’s set y = ceil(c/d). However note that at the beginning, we may have fewer than y operations actively running. So set y = min(ceil(c/d), i+1). i+1 is simply the #max of operations that could possibly have a chance of extending outwards. Then the #of extra units of time will be as follows:
c * y - [d*(y-1)*y*/2]. Thus their total extra contribution will be the previous expression multiplied by b. Subtract the new expression from our earlier sum and we get the total sum of all damage dealt to the monster.