Create D. Grime Zoo
https://codeforces.com/contest/1411/problem/D

Algorithm: Let’s look at a consecutive pair of question marks. Let’s say that the left question mark is turned into a 1 and the right question mark is turned into a 0. Then it’s actually always optimal to swap them if x <= y. You can prove this by doing the math based on the #of 0’s and 1’s between the 2 question marks. Note that none of the subsequences which have at least one element outside the inclusive interval of the consecutive pair of question marks are affected. Thus if x <= y, it’s always optimal for some prefix of the question marks to be 0’s and for the remaining suffix to be 1’s. If x > y, then we can just swap x and y and replace every ‘0’ with a ‘1’ and every ‘1’ with a ‘0’ in the test data and continue using the first idea. Implementing this idea quickly is described in the code