Create E. Decryption
https://codeforces.com/contest/1419/problem/E

Solution:

Note that since any number has < 1e9 has ~1000 divisors we can just take the prime factorization and recursively generate all factors > 1.

Let’s get the prime factorization of the number. Then if the number has exactly 1 prime factor, then just list all of its divisors and the minimum # of needed moves is just 0. If the number has exactly 2 prime factors, then we have 2 cases. First of all n = a * b. In this case, the best we can do is the following: a b a*b. The minimum # of needed moves in this case is 1. Otherwise, we can use the following arrangement: all multiples of a, a*b, all multiples of b, n. This will work and will need 0 extra moves. Note that when adding all multiples of a, don’t include a*b or n. Similarly when adding the multiples of b, don’t include a*b and n. What you can do is consider the lowest prime divisor of each factor of the original number and use that lowest prime divisor to group where each factor will belong. Now if the number has 3+ prime factors, then here is a construction that will work with 0 additional moves. Let n = p1^a1 * p2^a2 * p3^a3 * … pk^ak.

p1*p2, all multiples of p2, p2*p3, all multiples of p3, … p{k-1}*pk, all multiples of pk, pk*p1, all multiples of p1.

Once again when adding multiples of a prime, only use the factors for which that prime is the lowest prime that divides it. Also avoid using the connecting elements that connect different groups of multiples.

I’m not going to comment the code b/c it’s super messy and all you need to solve this problem is the construction