Create G. Reducing Delivery Cost
https://codeforces.com/contest/1433/problem/G

Solution: Let’s first run dijakstra’s algorithm n times in order to find the smallest distance between every pair of nodes. Then we can just sum the costs of the shortest paths for each route to get the answer if we don’t change any edge. Otherwise if we change on edge, let’s brute force this edge. For each edge, we’ll temporarily change it’s value to 0. Then we’ll run 2 dijakstra’s from either node connected by the edge. Let’s let the edges of the node be a and b respectively. Then for each route, the shortest distance between the 2 nodes in the route is either their initial precalculated distance, or the distance if one of the nodes in the route goes to one of the nodes of the 0-cost edge and the other node in the route goes to the other node in the edge. We sum this over all routes to get the total cost for changing the current edge to a 0. Do this process for all edges and use the minimal answer.