Create E. Xor Tree
https://codeforces.com/contest/1447/problem/E

Observation: First sort the numbers. No let’s first look at the MSB. Some prefix of the numbers will have a 0 as their bit at this position while the remaining suffix will have a 1 as their bit at this position. If the prefix or suffix is sized 0, then we’ll decrement the current bit we’re looking at by 1 and look at the same range again. Otherwise we’ll recursively find the best answer for the prefix with the current bit minus 1 and the best answer for the suffix with the current bit minus 1 and combine those answers to get the best answer for the current range. This works b/c the #’s are in sorted order and the way we recursively split our ranges ensures that for the current bit, the prefix of numbers in the range will always have a 0 at that position while the remaining suffix of numbers in the range will always have a 1 at that position. Our formula is as follows: Let f(a, b) = min # of deletions needed to ensure that the numbers from indices a to b are a tree. Then if x is the largest index in [a, b] such that the number at index x has a 0 in the place of the current bit, then our transition is as follows:

f(a, b) = min(x-a+1 + f(x+1, b), b-(x+1)+1 + f(a, x))

This is b/c every number in the prefix will have its edges directed to each other while every number in the suffix will have its edges directed to each other. Thus we’ll need to pick either the prefix or the suffix to continue using and deleting nodes from AND make sure that the remaining group only has 1 number. If the remaining group has only 1 number, then it will obviously be forced to have its edge connected to the dominant group that we picked, thus ensuring that our graph is a tree.

So just implement a straightforward recursion

Additionally there are a few base cases. If the length of the current range is 1 or 2, then the answer is 0 since the nodes will have to form an edge with each other and form a tree.