Create F. Array Partition
https://codeforces.com/contest/1454/problem/F

Solution:
First coordinate compress all the values in the array. Now let’s iterate over the size of the prefix. Let’s say that the maximum value in the current prefix is x. Then we’ll use binary search to find the first location of x after the current prefix. This will be the end of our middle section. Finally we’ll use the last location of x as the start of final section. Note that there may be some extra space between the second segment and the final segment. For each index, we’ll store the first index to its right with a smaller value and the first index to its left with a smaller value. This will help us find the first index that the middle segment can’t expand to and the first index the final segment can’t expand to. Split into cases and construct a valid solution for the sizes of the middle and final section. Note that there is ALOT of implementation and you need to check a ton of cases and also verify that the subsegments you’re iterating on even have the necessary minimums and maximums to begin with