Create F. New Year's Puzzle
https://codeforces.com/contest/1472/problem/F

Solution:

Let’s first find a solution for n <= 1e5. The optimal greedy solution is as follows. If the current column doesn’t have any blocked cells, then it’s always optimal to place a vertical domino here. This is b/c placing 2 horizontal dominos instead may trap you and force you into a suboptimal solution. Otherwise if the current column has 2 blocked cells ignore it. Otherwise if the current column has exactly 1 cell, place a horizontal domino. If any contradictions occur, then there isn’t a solution.

Now let’s optimize this idea for n <= 1e9. Let’s sort all the columns with at least 1 blocked cell. We’ll also add n+1 as a column with 2 blocked cells since you can’t go past n. If the current column is completely blocked, move on. Otherwise if the current column has 1 blocked cell, then compute the distance between the current column and the next column with at least 1 blocked cell. If this next column is completely blocked, there’s no solution. Otherwise if the same row is blocked on this next column as well, then we have a solution if the distance is odd. If the opposite row is blocked on this next column as well, then we have a solution if the distance is odd. You can see this all from examples. So just simulate this process