Create B1. Painting the Array I
https://codeforces.com/contest/1479/problem/B1

Solution:

Let’s build the 2 subsequences element by element. Note that all we care about is the total answer and the latest element in the 2 subsequences. Now we’ll determine which subsequence to place the current element in. If the current element is equal to the latest element in both of the subsequences, then it doesn’t matter which subsequence we place the current element in since the answer won’t change. If the current element is equal to the latest element in EXACTLY one of the subsequences, then place the current element in the other subsequence to increase the answer by 1. Also update the latest element of this subsequence. If the current element is different from the latest element in both subsequences, then determine for each latest element in the subsequence the next index it appears in. Whichever latest element appears sooner is the one we’ll place our current element in in order to maximize the answer