Create B2. Painting the Array II
https://codeforces.com/contest/1479/problem/B2

Solution:

Let’s build the 2 subsequences element by element. Note that all we care about is the total answer and the latest element in the 2 subsequences. Now we’ll determine which subsequence to place the current element in. If the current element is equal to the latest element in one of the subsequences, then place it in that subsequence in order to minimize the answer. Otherwise if one of the subsequences is empty while the other subsequence’s latest element isn’t equal to the current element, place the current element in the empty subsequence, increase the answer, and update the latest element of the empty sequence. Otherwise if the latest element in both subsequences are different from the current element, then determine the next closest index that the latest element in each subsequence appears in. Whichever one appears in a farther index will be the subsequence where we place our current element in.