Create D. Max Median
https://codeforces.com/contest/1486/problem/D

Solution: Let’s binary search on the answer. Let’s try to determine if a median >= x is possible. Let’s rewrite the array. If an element is < x, replace it with a -1. Otherwise if an element is >= x, replace it with a 1. Then if we can find a subarray with length >= k with a positive sum, then we’ve determined that a median >= x is possible. This second part is doable by computing prefix sums and maintaining at each index the minimum of all prefix sums up to that point